preserve the pre-skia-update C API for some simple cases
This doesn't cover all the API changes, just the most trivial
cases for now.

the new, incompatible versions of functions were renamed to *_v2
unless otherwise noted

all-zero sk_sampling_options_t seems to work as a reasonable default,
so in some places where it wasn't passed before that's what we will
pass to the C++ API

sk_filter_quality_t seems to have no trivial replacement, so it will
be ignored.

the canvas functions which were changed to matrix44 seemingly didn't
need to be, the new ones were renamed to *_m44

enums and structs which had entries added in the middle  were reordered
so that the new additions are at the end, in order to not change the values
assigned to the existing entries.